+++ src/main/java/com/trabalho/nexus/metafinanceira/MetaFinanceiraResponseDTO.java
@@ -5,13 +5,12 @@ import java.time.Instant;
 public record MetaFinanceiraResponseDTO (
 	    Long id,
 	    String descricao,
-	    Long valorMeta, 
+	    Double valorMeta, 
 	    Double saldo, 
 	    Instant dataInicial, 
 	    Instant dataFinal, 
 	    Long usuarioId,
 	    Character status,
 	    double percentualRendimento,
-	    
 	    String tipoInvestimento
 	) {}
